# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Capyba",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
        "/auth": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Loga o usuário",
                "description": "Loga o usuário e retorna o token de autenticação",
                "operationId": "0fa80934886d2117ad53d0a4761476a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário logado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "E-mail e ou senha inválidos."
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Desloga o usuário autenticado",
                "description": "Desloga o usuário autenticado",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/password-forgot": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Envia o link de redefinição de senha",
                "description": "Envia um link de redefinição de senha para o e-mail do usuário.",
                "operationId": "b268a129c129a778ae5b9e0592ba5ebc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Link de redefinição enviado com sucesso"
                    },
                    "400": {
                        "description": "Erro ao enviar o link"
                    }
                }
            }
        },
        "/auth/password-reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Redefine a senha do usuário",
                "description": "Redefine a senha do usuário utilizando o token de redefinição.",
                "operationId": "663262e7a35f18f82d5eba60842a689c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha redefinida com sucesso"
                    },
                    "400": {
                        "description": "Erro ao redefinir a senha"
                    }
                }
            }
        },
        "/user/new-password": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Atualiza a senha do usuario",
                "description": "Atualiza a senha do usuario.",
                "operationId": "1b12d4af39f7163eed409267f5254a80",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "newPassword"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    },
                                    "newPassword": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha atualizada com sucesso"
                    },
                    "400": {
                        "description": "Erro ao atualizar a senha"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use o token fornecido pelo Sanctum. Exemplo: 'Bearer {token}'",
                "scheme": "bearer"
            }
        }
    },
...
}